Seperate server functionality and add tests for GET requests
